Refactor consumer and create handler abstraction
Make the handling of kafka messages more flexible, by removing the need
of extending the Consumer base class. The user should only provide a
callable that receives the raw (or decoded) message as argument.

Create the ConsumerBuilder class and update the examples to use the
builder.

Create abstraction ConsumerListener with the goal of removing logging
concerns from this library, delegating this for the clients.

Create the MessageDecoder. This will allows us to add support for
different types of messages encoding in kafka: avro schemas, simple json
messages, etc. Making it possible to create reusable decoders.