Refactor consumer listener contract and improve ci (#11)
The consumer listener received as argument for the messageDecodingFailed
event the raw message. However, this may not be enough for the clients
to take action on it, so now we provide the exception object thrown
which also contains the raw message.

Improve CI pipeline by running checks with the whole range of defined
dependencies.